refactor(foundation): move threat_modeling + rule data types (A16, A17)
Two related tactical extractions that unblock another 117 deprecation
warnings in mockforge-http (468 → 351):

**A16 — threat_modeling types** (`foundation::threat_modeling_types`):
- `ThreatAssessment`, `AggregationLevel`, `ThreatLevel`, `ThreatCategory`,
  `ThreatFinding`, `RemediationSuggestion`, `ThreatModelingConfig`
- All pure data (no OpenApiSpec / LLM deps); `ThreatAnalyzer` impl stays
  in core because it depends on `OpenApiSpec` and LLM integration.
- Core's `contract_drift::threat_modeling::types` re-exports from
  foundation. http handlers for `threat_modeling.rs`,
  `contract_health.rs`, and the router wiring now import data types
  from foundation directly.

**A17 — rule_generator data types**
(`foundation::intelligent_behavior::rule_types`):
- `ExamplePair`, `ErrorExample`, `PaginatedResponse`, `CrudExample`,
  `ValidationRule`, `PaginationRule`, `RuleType`, `PatternMatch`,
  `RuleExplanation`
- `RuleGenerator` impl (uses `LlmClient`) stays in core.
- Core's `rule_generator.rs` re-exports from foundation so existing
  callers keep working. http's `management/mod.rs` and `ai_gen.rs`
  import `RuleExplanation` and `RuleType` from foundation directly.

All 574 http + 1818 core + 42 foundation tests pass; full workspace
clippy clean.
